requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.10
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
black==24.10.0
boto3==1.37.14
botocore==1.37.14
bs4==0.0.2
celery==5.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
decorator==5.2.1
dill==0.4.0
Django==5.1.4
django-celery-beat==2.8.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-polymorphic==3.1.0
django-ses==4.4.0
django-storages==1.14.5
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
executing==2.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==5.13.2
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.6
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.11.5
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tomlkit==0.13.3
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
